requirements.txt
discord.py
yt-dlp
PyNaCl
python-dotenv
discord.py[voice]
aiosqlite
aiohttp
cryptography
flask

## FFmpeg 설치해줘야 함

## bash -> winget install ffempeg